moved styles and Options components into there own files
