+++ package.json
@@ -78,6 +78,7 @@
     "typescript": "^3.9.7"
   },
   "dependencies": {
+    "caniuse-lite": "^1.0.30001187",
     "lodash.get": "^4.4.2"
   }
 }
