+++ src/components/index.ts
@@ -1,4 +1,4 @@
-export * from './LocalePickerToolbarPlugin';
+export * from './toolbarPicker';
 export * from './InlineTranslatedText';
 export * from './ModalBuilder';
 export * from './I81nProvider';
